Refactor: Extract mock data and minor code improvements
- Extract mock data to dedicated MockDataService for better reusability and maintainability
- Implement MARK comments for improved code navigation and readability
- Adjust minor UI setup methods for consistency across the project
