+++ Tracker/SceneDelegate.swift
@@ -16,7 +16,6 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         guard let windowScene = (scene as? UIWindowScene) else { return }
         window = UIWindow(windowScene: windowScene)
         window?.rootViewController = MainTabBarController()
-//        window?.rootViewController = SelectTrackerViewController()
         window?.makeKeyAndVisible()
     }
